The class in main.dart was splited in new class
Hey Rashaad! Dividing the class into different files is a good practice. Improving your code reading and maintenance. ;D
