+++ lib/main.dart
@@ -1,25 +1,9 @@
 import 'package:flutter/material.dart';
 
+import 'home_page.dart';
+
 void main() {
   runApp(MyApp());
-}
-
-class Post {
-  String body;
-  String author;
-  int likes = 0;
-  bool userLiked = false;
-
-  Post(this.body, this.author);
-
-  void likePost() {
-    this.userLiked = !this.userLiked;
-    if (this.userLiked) {
-      this.likes += 1;
-    } else {
-      this.likes -= 1;
-    }
-  }
 }
 
 class MyApp extends StatelessWidget {
@@ -33,106 +17,6 @@ class MyApp extends StatelessWidget {
         visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
       home: MyHomePage(),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  List<Post> posts = [];
-
-  void newPost(String text) {
-    this.setState(() {
-      posts.add(new Post(text, "Rash"));
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(title: Text('Hello World!')),
-        body: Column(children: <Widget>[
-          Expanded(child: PostList(this.posts)),
-          Expanded(child: TextInputWidget(this.newPost))
-        ]));
-  }
-}
-
-class TextInputWidget extends StatefulWidget {
-  final Function(String) callback;
-  TextInputWidget(this.callback);
-
-  @override
-  _TextInputWidgetState createState() => _TextInputWidgetState();
-}
-
-class _TextInputWidgetState extends State<TextInputWidget> {
-  final controller = TextEditingController();
-
-  @override
-  void dispose() {
-    super.dispose();
-    controller.dispose();
-  }
-
-  void click() {
-    widget.callback(controller.text);
-    controller.clear();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return TextField(
-        controller: this.controller,
-        decoration: InputDecoration(
-            prefixIcon: Icon(Icons.message),
-            labelText: "Type a message",
-            suffixIcon: IconButton(
-              icon: Icon(Icons.send),
-              splashColor: Colors.blue,
-              tooltip: "Post message",
-              onPressed: this.click,
-            )));
-  }
-}
-
-class PostList extends StatefulWidget {
-  final List<Post> listItems;
-
-  PostList(this.listItems);
-
-  @override
-  _PostListState createState() => _PostListState();
-}
-
-class _PostListState extends State<PostList> {
-  @override
-  Widget build(BuildContext context) {
-    return ListView.builder(
-      itemCount: this.widget.listItems.length,
-      itemBuilder: (context, index) {
-        var post = this.widget.listItems[index];
-        return Card(
-            child: Row(children: <Widget>[
-          Expanded(
-              child: ListTile(
-            title: Text(post.body),
-            subtitle: Text(post.author),
-          )),
-          Row(
-            children: <Widget>[
-              IconButton(
-                icon: Icon(Icons.thumb_up),
-                onPressed: post.likePost,
-              )
-            ],
-          )
-        ]));
-      },
     );
   }
 }
